+++ NN_basic/Network.h
@@ -8,7 +8,7 @@ class Network
 {
 public:
 	Network(int, int, int);
-	float* process(float*);
+	double* process(double*);
 private:
 	int input_num;
 	int hidden_num;
@@ -16,9 +16,9 @@ private:
 	InputNeuron* input_neurons;
 	HiddenNeuron* hidden_neurons;
 	OutputNeuron* output_neurons;
-	float* in_res;
-	float* hid_res;
-	float* out_res;
+	double* in_res;
+	double* hid_res;
+	double* out_res;
 
-	float learning_rate;
+	double learning_rate;
 };
